requirements.txt
openai>=1.40.0
gradio>=4.44.0
python-dotenv>=1.0.1
youtube-transcript-api>=0.6.1
requests>=2.32.3
networkx>=3.3
plotly>=5.24.0
spacy>=3.7.4
